Add highlighting to REPL output
REPL output from Cryptol frequently contains source locations. With
these changes, the source locations are recognized and annotated with
an overlay that causes them to be clickable links to the position in
question.

Additionally, the strings "[error]" and "[warning]" are highlighted
with the corresponding faces from compilation-mode, so that they are
easier to notice.